+++ controllers/discord.js
@@ -1,25 +1,32 @@
-import got from "got";
-import Error from "./error.js";
+import Controller from "./template.js";
 
-export default class Discord {
-	static webhook = null;
+export default class Discord extends Controller {
+	#active = false;
+	#token = null;
 
-	constructor (data) {
-		if (typeof data.webhook !== "string") {
-			throw new Error({ message: "Discord webhook must be a string" });
-		}
+	constructor (config) {
+		super();
 
-		Discord.webhook = data.webhook;
-	}
-
-	async send (data, options = {}) {
-		if (!Discord.webhookAuth) {
+		if (config.enabled === false) {
 			return;
 		}
-        
-		const embed = options.skipEmbed ? data : Discord.generateEmbed(data);
-		const res = await got({
-			url: Discord.webhookAuth,
+
+		const token = config.webhook;
+		if (!token) {
+			throw new sr.Error({ message: "Discord webhook doesn't exist" });
+		}
+		
+		this.#token = token;
+		this.#active = true;
+	}
+
+	async send (embed) {
+		if (!this.#active) {
+			throw new sr.Error({ message: "Discord is not active" });
+		}
+		
+		const res = await sr.Got({
+			url: this.#token,
 			method: "POST",
 			responseType: "json",
 			json: {
@@ -30,19 +37,18 @@ export default class Discord {
 		});
 
 		if (res.statusCode !== 204) {
-			throw new Error({
-				message: "Error when sending message to Discord",
+			throw new sr.Error({
+				message: "Failed to send message to Discord",
 				args: {
 					statusCode: res.statusCode,
+					statusMessage: res.statusMessage,
 					body: res.body
 				}
 			});
 		}
-
-		return true;
 	}
 
-	static generateEmbed (messages) {
+	generateEmbed (messages) {
 		let message = "";
 
 		for (const data of messages) {
@@ -57,13 +63,13 @@ export default class Discord {
 		}
         
 		return {
+			color: 0xBB0BB5,
 			title: "Honkai: Star Rail Auto Check-in",
 			author: {
 				name: "Honkai: Star Rail",
 				icon_url: "https://i.imgur.com/o0hyhmw.png"
 			},
 			description: message,
-			color: 0xBB0BB5,
 			timestamp: new Date(),
 			footer: {
 				text: "Honkai: Star Rail Auto Check-in"
@@ -71,13 +77,5 @@ export default class Discord {
 		};
 	}
 
-	static get webhookAuth () {
-		const auth = Discord.webhook;
-
-		if (!auth || typeof auth !== "string") {
-			return null;
-		}
-
-		return auth;
-	}
+	get active () { return this.#active; }
 }