+++ controllers/template.js
@@ -1,0 +1,6 @@
+/* eslint-disable no-unused-vars */
+export default class Controller {
+	async send (message, options = {}) {}
+
+	destroy () {}
+}